- Added new DeviceLocation class which uses the FusedLocationProvider - LocationScene: added removeLocationNode method which remove a LocationNode from the screen
